requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.24.5
idna==3.7
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sniffio==1.3.1
starlette==0.37.2
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
